WIP: Update kpt to use cli-utils with custom inventory storage logic
This commit point kpt to use a forked version of cli-utils that includes
the Apply and ApplyWithPrune interface methods to allow custom logic of
applying inventory on the live cluster by the client of cli-utils. This
is pre-requisite work to enable auto-sharding of ResourceGroup CRs by
kpt.

Do not merge until cli-utils is updated to include the forked changes.